Title: 합과 차의 나눗셈, Time: 93ms, Memory: 9MB, Status: Wrong Answer - CodeTree

+++ 241016/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
-        double a = 10;
-        double b = 3;
+        int a = 10;
+        int b = 3;
         double c = a+b;
         double d = a-b;
